discovery-engine/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
anthropic==0.40.0
python-dotenv==1.0.1
PyMuPDF==1.24.14
python-docx==1.1.2
neo4j==5.27.0
httpx==0.27.2
